+++ test/runtests.jl
@@ -143,15 +143,13 @@ let
     @test getMean(z[5]) == 5
     @test getVar(z[5]) == 1
 
-    c = (3v+1)*x + 10 <= 20
-    @test conToStr(c) == "(3 v + 1)*x + -10 <= 0"
-    addConstraint(m, c, with_probability=0.05)
+    @addConstraint(m, (3v+1)*x + 10 <= 20, with_probability=0.05)
     @test conToStr(JuMPChance.getCCData(m).chanceconstr[1]) == "(3 v + 1)*x + -10 >= 0, with probability 0.95"
     @addConstraint(m, (3v+1)*x + 10 <= 20, with_probability=0.95)
     @test conToStr(JuMPChance.getCCData(m).chanceconstr[end]) == "(3 v + 1)*x + -10 <= 0, with probability 0.95"
-    @test_throws ErrorException addConstraint(m, c, with_probability=10.0)
-    c = (3v+1)*x + 10 >= 20
-    @test conToStr(c) == "(3 v + 1)*x + -10 >= 0"
+    @test_throws ErrorException @addConstraint(m, (3v+1)*x + 10 >= 20, with_probability=10.0)
+    @addConstraint(m, (3v+1)*x + 10 >= 20, with_probability=0.96)
+    @test conToStr(JuMPChance.getCCData(m).chanceconstr[end]) == "(3 v + 1)*x + -10 >= 0, with probability 0.96"
 
     ccaff = JuMPChance.CCAffExpr()
     ccaff.constant = 2v
@@ -207,6 +205,23 @@ let
         @test_approx_eq_eps getValue(z) -1/quantile(Normal(0,1),0.95) 1e-6
     end
 end
+
+# test deprecated operator overloads
+let
+    for method in [:Reformulate,:Cuts]
+        m = ChanceModel()
+        @defIndepNormal(m, x, mean=0, var=1)
+        @defVar(m, z >= -100) # so original problem is bounded
+
+        @setObjective(m, Min, z)
+        addConstraint(m, z*x >= -1, with_probability=0.95)
+
+        status = solve(m, method=method)
+        @test status == :Optimal
+        @test_approx_eq_eps getValue(z) -1/quantile(Normal(0,1),0.95) 1e-6
+    end
+end
+
 # flipped constraint sense
 let
     for method in [:Reformulate,:Cuts]